Worker: Rename terminationEvent and shutdownEvent on WorkerThread
In the current implementation, terminationEvent is signaled when the worker
thread is shut down and shutdownEvent is signaled when the worker thread is
requested to terminate. This would be confusing.

This CL just renames terminationEvent to shutdownEvent and shutdownEvent to
terminationEvent. The behavior should never be changed by this CL.

BUG=575532

Review-Url: https://codereview.chromium.org/1985943002
Cr-Original-Commit-Position: refs/heads/master@{#394705}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73364e66e7bba3522f28d5146d7e78ef8bb39f5a