[css tables] Don't pass table width increase due to % columns to parents
A table cell that specifies % width can increase the width of its table,
but shouldn't increase the width of any grandparent table cell.

This makes Blink match FF and Edge.

LayoutTable::layout calls LayoutTable::updateLogicalWidth, which
eventually calls into
TableLayoutAlgorithmAuto::computeIntrinsicLogicalWidths, which used to
do the percent scaling. This doesn't work because when the table is
inside a table-cell, LayoutTableCell calls the child table's
TableLayoutAlgorithmAuto::computeIntrinsicLogicalWidths, returning the
scaled width, not the intrinsic width.

This patch moves the scaling to a separate step that's only called in
LayoutTable::updateLogicalWidth, not in
TableLayoutAlgorithmAuto::computeIntrinsicLogicalWidths.

BUG=178369,427994

Review-Url: https://codereview.chromium.org/1921973005
Cr-Original-Commit-Position: refs/heads/master@{#394715}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9428cfb16993a2329e87c65da096ca295132ef0f