Include selection rect in paint invalidation for LayoutText.
This change is intended to fix the paint invalidation rect, and thus
the visual rect, as part of the linked bug http://crbug.com/605319
spun out of http://crbug.com/529938. There is more explanation of
the intended use of visual rects in my comment at
http://crbug.com/604883#c5

The visual rects that are incorrect without this change are not
currently used by live code, but will be used as part of
http://crrev.com/1484163002. This change fixes about 15 LayoutTests
with that patch applied, mainly under editing.

BUG=605319

Review-Url: https://codereview.chromium.org/1955423004
Cr-Original-Commit-Position: refs/heads/master@{#394807}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76f9a545cdf9903cded65e98b616a14976428aa8