ifdef Snappy's version of DISALLOW_COPY_AND_ASSIGN.
This avoids macro redefinition errors if we change Chromium's definition of this
macro to use "= delete".

Upstream rejected changing this macro in the same way because they want to stay
C++98-only, but agreed to accept this change (though as of right now that hasn't
actually happened yet).

BUG=447156
TEST=none

Review-Url: https://codereview.chromium.org/1997503005
Cr-Original-Commit-Position: refs/heads/master@{#394913}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ce70473fc0976e9916749a5f69dcabfea0f40c0