Disallow throttling while running requestAnimationFrame callbacks
If a requestAnimationFrame callback tries to do a synchronous style
computation in another frame which is currently throttled, this fails
because throttling is allowed for the entire BeginMainFrame, which
includes running rAF callbacks as a substep.

This patch fixes the issue by disallowing throttling while executing rAF
callbacks. Note that rAF callbacks for throttled frames are still
skipped.

BUG=607336

Review-Url: https://codereview.chromium.org/1994133002
Cr-Original-Commit-Position: refs/heads/master@{#395054}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: afdf51c967149db6bd8d6a533886efc957435e18