[css-grid] Horizontal scroll must account for grid container's height.
We don't consider the scrollbar when computing the grid container's
logical height at the beginning of the 'layoutBlock' logic. We must
do it because otherwise, contentLogicalHeight may return a negative
value.

We use the result of the contentLogicalHeight function to compute the
row tracks. This logic asserts we don't receive negative values, hence
there are cases which may violate such assertion.

BUG=583743

Review-Url: https://codereview.chromium.org/1988403002
Cr-Original-Commit-Position: refs/heads/master@{#395058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7966603b98ac3ca931ab9f04846d898c2242e944