Synchronously send EmulateNetworkConditions command when a new agent is attached to DevTools
When a new ServiceWorker is registered or an existing ServiceWorker registration starts updating, the new ServiceWorker is attached to the opened DevTools.
If the network throttling is set, the DevTools sends emulateNetworkConditions command to the browser process in the callback function of canEmulateNetworkConditions.
But at this moment, Runtime.run command has already been sent to the ServiceWorker.
The network request for the ServiceWorker script may be fetched before the browser process recieves the "emulateNetworkConditions" command.
So the network throttling doesn't work for the ServiceWorker's script request.

WebInspector.ServiceWorker()
|- WebInspector.targetManager.createTarget()
|   |- WebInspector.TargetManager.targetAdded()
|      |- WebInspector.MultitargetNetworkManager.targetAdded()
|         |- networkAgent.canEmulateNetworkConditions(callback)
|            ==> send "canEmulateNetworkConditions" command
|              ==> send "emulateNetworkConditions" command in the callback
|- runtimeAgent().run()
     ==> "Runtime.run()"

To solve this problem, this CL change NetworkManager to Synchronously send EmulateNetworkConditions command when a new agent is attached to DevTools if the network throttling is set.

BUG=588584

Review-Url: https://codereview.chromium.org/1998573002
Cr-Original-Commit-Position: refs/heads/master@{#395082}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2594a07d804d63a11ab847d5d76ed322e827a3e7