MediaCaptureFromElement: add support for audio captureStream().
This CL extends support for capturing the audio part of
a <video> or <audio> tags ( "capture" here means creating
a MediaStream out of the HTMLElement)

It introduces an HtmlAudioCapturerSource is-a AudioCapturerSource
wrapped into an ExternalMediaStreamAudioSource to produce data
towards the audio track.

HtmlAudioCapturerSource also plugs into the
WebMediaPlayer's WebAudioSourceProviderImpl to get
a copy of the audio being rendered.

Unit tests are added, and the existing LayouTests
revamped (and split into several files for clarity).

BUG=569976, 575492

TEST= run chromium with
 --enable-blink-features=MediaCaptureFromVideo
 against e.g.
https://rawgit.com/Miguelao/demos/master/videoelementcapture.html

Review-Url: https://codereview.chromium.org/1599533003
Cr-Original-Commit-Position: refs/heads/master@{#395205}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77d0d446e58afbf7fab215113fcf9fe9c97e94e3