GYP: In kasko target, export dependent settings of kasko_features
Many targets depend on the kasko target for a dependency on the generated
header kasko_features.h. However, when kasko is not enabled, the kasko
target was not exporting the fact that kasko_features is a hard
dependency, causing flaky build failures. This CL fixes the bug.

BUG=611319
TBR=pmonette

Review-Url: https://codereview.chromium.org/2007513002
Cr-Original-Commit-Position: refs/heads/master@{#395314}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af480cf00ad993ac39e9e3253701879b17d2140b