OOPIF: Replicate allowFullscreen flag.
A frame is allowed enter fullscreen if the allowFullscreen attribute
is present in all of its ancestor <iframe> elements.  For OOPIF, this
attribute needs to be replicated: when a parent frame updates
allowFullscreen for a remote child frame, the child's RenderFrame
needs to learn about this.  Additionally, in a case like
A-embed-B-embed-C, when A updates allowFullscreen for B, C's process
also needs to be notified, since this affects C's fullscreen
calculation.  Therefore, B's RenderFrameProxy in C's process also needs
to receive this update.

This CL replicates the allowFullscreen flag by adding it to
WebFrameOwnerProperties, which are already replicated to the child's
RenderFrame.  It also adds logic to replicate WebFrameOwnerProperties
to the child's proxies as well, so that it can handle allowFullscreen
calculation across multiple remote ancestors.

BUG=550497
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/1938753002
Cr-Original-Commit-Position: refs/heads/master@{#395496}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 788f45b374a8ad32e8a9a6eb910ded1360f33caf