[css-grid] Update <fixed-size> syntax
The syntax for <fixed-size> has been updated on the spec:
https://drafts.csswg.org/css-grid/#typedef-fixed-size

New syntax is:
  <fixed-size> =
    <fixed-breadth> |
    minmax( <fixed-breadth> , <track-breadth> ) |
    minmax( <inflexible-breadth> , <fixed-breadth> )

This means that it's enough to have one <fixed-breadth>,
it doesn't matter if it's as minimum or maximum.
Before it was required that the minimum was fixed.

Updated test to check the new expected behavior.

BUG=611412
TEST=fast/css-grid-layout/grid-element-auto-repeat-get-set.html

Review-Url: https://codereview.chromium.org/2001113002
Cr-Original-Commit-Position: refs/heads/master@{#395538}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e41b40417c0ba7a1f44a0cb6d636ffc76811d666