OwnPtr: Add reset() as an alias of clear().
We are planning to rename OwnPtr::clear() to reset(), as a part of
OwnPtr -> std::unique_ptr migration.

This patch adds a new member function reset() that is an alias of
clear() in order to facilitate the rename process. Eventually reset()
will take over clear().

BUG=611661

Review-Url: https://codereview.chromium.org/2004383002
Cr-Original-Commit-Position: refs/heads/master@{#395586}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e538ecb8f6819b5b39ec8146b9ae566f519a939