[SVG] Fix bounding box computation for transformed clips
Local clipPath transforms should be applied before the objectBoundingBox
matrix.  We follow this rule for painting, but not when computing the
clip bbox: the local matrix is fused with the local clip bounds and the
result is run through the OBB matrix => OBB is logically applied before
the local transform (T = [OBBMatrix] X [LocalMatrix]).

This means our clip bbox can be wildly inaccurate in the presence of
OBB + local transform (due to the large scale factors involved),
resulting in incorrect quick-rejects.

The CL splits the cached local clip bounds from the local transform,
and ensures that the result is computed using
[LocalMatrix] X [OBBMatrix].

R=fs@opera.com

Review-Url: https://codereview.chromium.org/2007633003
Cr-Original-Commit-Position: refs/heads/master@{#395601}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d87f7dfaab3c3b75d7be1a3fcc51f7466d4a3582