Fix whitespace handling in parseSimpleTransformList
The loop would break if there was any trailing whitespace after the
args of the transform operation function, but this meant that if you
put a space between the functions like 'translateX(1px) translateY(1px)'
you'd fall out of the fast path when hitting the space between the
two functions. Any trailing space would similarly fall out of the fast
path and result in wasting all of the work building the transform list
only to discard it.

This patch instead skips whitespace between the functions and also adds
a bunch of tests for the fast paths.

BUG=606211

Review-Url: https://codereview.chromium.org/2007823002
Cr-Original-Commit-Position: refs/heads/master@{#395610}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f88403bf3e53a955cee48e172269d0ed94f2595