Track whether origin trial feature bindings have been installed
This patch adds to the OriginTrialContext a set of features for which V8
bindings have been installed, and two methods to maintain this set.

In a subsequent CL, we will be changing the origin trial V8 bindings to be installed conditionally after the V8 context is initialized. At that point, these methods are going to be used to track which features have already been added, so that we don't attempt to install any feature twice.
BUG=584367

Review-Url: https://codereview.chromium.org/2003603002
Cr-Original-Commit-Position: refs/heads/master@{#395704}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8300471d75ffba2c2234e9424a6a68798f5c90e