Worker: Make WorkerMicrotaskRunner a private inner class of WorkerThread
This CL makes WorkerMicrotaskRunner a private inner class of WorkerThread
because the task runner is basically used only from WorkerThread and it
needs to access a function of WorkerThread that is currently public but should
be private.

This also renames performShutdownTask() to performShutdown() for improving
naming consistency with prepareForShutdown().

This should not change any behavior.

BUG=487050, 575532

Review-Url: https://codereview.chromium.org/2006273003
Cr-Original-Commit-Position: refs/heads/master@{#395817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 890dd5caf4fd966ad59b219b53c9141b4a7399a1