[css-grid] Refactor populateGridPositions()
LayoutGrid::populateGridPositions() was doing exactly the same thing for
columns and rows but using different data structures. That lead to a lot of
duplicated code. It's easy to refactor it in a new function that properly
select the data structures to operate on based on the direction.

No new tests required as there is no change in behaviour.

Review-Url: https://codereview.chromium.org/2011763004
Cr-Original-Commit-Position: refs/heads/master@{#395912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f963d8f58658ad28cdeaf6033eefbd8d51a6246f