Throw exception for non-finite values in setValueCurve
If the curve given to setValueCurveAtTime contains non-finite values,
throw an exception.  All other AudioParam methods including the value
setter or the automation methods do not allow a non-finite value to be
specified.  The same should be true for the curve values given to
setValueCurveAtTime.

BUG=612787
TEST=audioparam-setValueCurve-exceptions.html

Review-Url: https://codereview.chromium.org/1995583002
Cr-Original-Commit-Position: refs/heads/master@{#396042}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e1870e6504d085c7b624ae16f55d03e46dc8ef0