Correctly position abspos content inside relpos inlines [spv2]
This patch fixes a class of bugs where absolutely positioned content
in a relative inline container would be positioned relative to the
origin instead of relative to the inline. This largely mirrors the logic
in PaintLayer::updateLayerPosition which used
LayoutInline::offsetForInFlowPositionedInline to position PaintLayers.

BUG=614257

Review-Url: https://codereview.chromium.org/2000423009
Cr-Original-Commit-Position: refs/heads/master@{#396324}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0079320b1c8fdad99ef3cb31b92ef09da6ecd2e0