Implement cookie-averse behavior for Suborigins
The draft of the Suborigin spec calls for Document objects in a
non-empty Suborigin to be cookie-averse, as defined by the HTML5.1 spec.
This implements that behavior by causing such Documents to return the
empty string when document.cookie is referenced, and to no-op when
document.cookie is set.

BUG=336894
R=mkwst@chromium.org

Review-Url: https://codereview.chromium.org/2020513002
Cr-Original-Commit-Position: refs/heads/master@{#396623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f7b7068772383fd88c0c54f044b28d9a52be8a3