Apply vertical-align style of <sub> and <sup> to child elements.
If the selected element has <sub> or <sup> ancestor element,
apply the corresponding style(vertical-align) to it so that
document.queryCommandState() works fine on the selected element.

<sub> and <sup> tags are represented with CSS vertical-align
property but they are not inherited. Therefore, we need to apply the
property to child elements because <sub> and <sup> tag can have
child elements such as <i> or <b> tag.

BUG=582225
TEST=editing/execCommand/queryCommandState-04.html
     editing/execCommand/queryCommandState-04-mac.html

Review-Url: https://codereview.chromium.org/1986563002
Cr-Original-Commit-Position: refs/heads/master@{#396763}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dfe3d34820670521568ca7dfe681f701da94325b