Add a stack of queues of elements with reaction queues.
The ordering of custom element 'reactions' (various kinds of
callbacks) is determined by a stack which is pushed and popped within
the scope of an operation. The stack entry is a queue of involved
elements; this is so callbacks occur in order. However each element
gets callbacks in an internally consistent chronological order. To
effect this, each element has an associated reaction queue.

BUG=594918

Review-Url: https://codereview.chromium.org/2027513002
Cr-Original-Commit-Position: refs/heads/master@{#396806}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e1a83859dd7df828f87fac61886f07f454dfffeb