Use the same policy to check paint rect for all cell paintings
Because we use object-based invalidation, if an object is not
invalidated, it must draw all or none. For a table cell, it draws
background from container, collapsed borders and own contents from
different callers. We should use the same policy to check paint
rect intersection to ensure all-or-none painting.

BUG=614233

Review-Url: https://codereview.chromium.org/2022653002
Cr-Original-Commit-Position: refs/heads/master@{#396856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e2f76f709bf856bbece8d45349cc7e2b1665333